Created new class for recommend functions
Removed the getRecommend, getUnder20, and getRandomProduct functions from the OnlineUser class and moved them to a new one.
